+++ Anarchy/ChangeLog.md
@@ -1,3 +1,5 @@
-﻿## Patch V1.4.2
-* Reinstate Unified Icon Dependency
-* Only loads English localization if file not found.
+﻿## Patch V1.4.3
+* Fixed tooltip related issues.
+* Updated localization. Added I18n Everywhere dependency.
+* Fixed Flaming chirper with disabled radio.
+* Added Opt-In option to automatically disable anarchy toggle while brushing objects.